Add logic to propagate CE-cancelation via channels
Adds a `CancelTemplate` for users to specify a method that should
be used for cancelation purposes. This LSP method is implemented
automatically. When invoked, the server looks up the list of running
calls it contains, and cancels the corresponding one.

On the client side, stubs are automatically wired so that canceling a
request in flight leads to a cancelation request propagating to the
server with the corresponding CallId.